server/requirements.txt
python-dotenv==1.0.0
pika==1.3.2
async-timeout==4.0.3
redis==5.0.4
torchaudio==2.0.2
torch==2.0.1
transformers==4.31.0
pydantic==2.5.2
pydantic_core==2.14.5